roachtest: split perturbation/* tests
This commit splits all the individual perturbations into separate files.
This will make it easier to find and update a perturbation.

Epic: none

Release note: None
